codes/Evaluation/requirements.txt
pandas==1.5.3
openai==0.27.2
langchain==0.0.208
langchain-community==0.1.3
langchain-openai==0.1.0
swarm==0.2.0
google-generativeai==0.1.0
